requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

bcrypt==4.1.2
beanie==1.25.0
discord.py==2.4.0
fastapi==0.109.0
motor==3.3.2
passlib==1.7.4
prettytable==3.11.0
pydantic==2.5.3
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==1.0.0
setuptools==75.1.0
tabulate==0.9.0
uvicorn==0.26.0
python-multipart==0.0.12
googletrans==4.0.0-rc1
